requirements.txt
fastapi
uvicorn
telethon
motor
python-jose[cryptography]
passlib[argon2]
pydantic
python-dotenv
cryptography
